Ensure calling destroy() works properly
It seems that Meteor deletes the ‘fa.editable’ jQuery data attribute on
the Froala Editor plugin instance by the time the
`Template.froalaReactive.destroyed` callback is invoked.  This meant
that the attempt to call FE’s `destroy()` method keep failing, because
the gateway function in FE could not get a reference to its FE object.

Workaround is to stuff a reference to the ‘fa.editable’ jQuery data
object into the froalaReactive template instance object, and then to
set the value of ‘fa.editable’ again in the destroyed callback, just
before the call to the destroy() method.  Seems to do the trick.  Using
the browser’s DOM inspector, I can now see that the ‘froala-editor’,
‘froala-modal’ and ‘froala-overlay’ divs are being removed from the DOM
when I navigate away from the page with the frail Editor instances on
it. Before this fix, they remained in the DOM.

Fixes #2